fix page redirecion error
