+++ src/App.js
@@ -1,10 +1,5 @@
-import axios from "axios";
-import { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
 import "./App.css";
-import { logoutHandler, settingInitialValues } from "./Containers/Redux/UserAuth";
-import { AuthenticatedRoutes } from "./Containers/Routes/Authenticated";
-import { UnAuthenticatedRoutes } from "./Containers/Routes/UnAuthenticated";
+import AppRoutes from "./Containers/Routes/AppRoutes";
 import { Buffer } from "buffer";
 import { useEffect } from "react";
 import { Suspense } from "react";
@@ -12,13 +7,6 @@ import { Spinner } from "./Components/Spinner/Loader";
 
 
 function App() {
-
-  const dispatch = useDispatch();
-  const [isLoading, setIsLoading] = useState(false)
-  let userFound = useSelector((state) => state.UserAuth.isLoggedIn);
-  let reduxToken = useSelector((state) => state.UserAuth.token);
-  const token = localStorage.getItem("token");
-
 
 //  function isTokenExpired(token) {
 //     const payloadBase64 = token.split(".")[1];
@@ -29,49 +17,10 @@ function App() {
 //     return expired;
 //   }
 
-  const ValidateToken = async (currToken) => {
-   
-    const resp = await axios({
-      method: "POST",
-      url: "https://plugin-nodejs-server.herokuapp.com/api/login",
-      headers: {
-        authorization: `Bearer ${currToken}`,
-      },
-    })
-      .then((res) => {
-        setIsLoading(false)
-        console.log("Logged In")
-
-        const checkExpiry = true;
-        if(checkExpiry){
-          console.log("Time not expired")
-          dispatch(settingInitialValues({ userToken:token , userLoggedIn :true}))
-        }else{
-          console.log("Time expired")
-          dispatch(logoutHandler())
-        }
-      })
-      .catch((er) => {
-       
-        console.log("Error in Validation", er)
-        dispatch(logoutHandler());
-      });
-  };
-
-
-    if (!token) {
-      dispatch(logoutHandler);
-    } else {
-    
-      ValidateToken(token);
-    }
-
-
-
   return (
     <Suspense fallback={<Spinner color='#2285b6'/>}>
     <div className="App">
-      {userFound ? <AuthenticatedRoutes /> : <UnAuthenticatedRoutes />}
+     <AppRoutes />
     </div>
     </Suspense>
   );
